+++ pyabc/cv/__init__.py
@@ -1,0 +1,1 @@
+from .bootstrap import calc_cv